release

+++ src/main/java/ru/tagall/emulatesystem/application/session/model/SessionDto.java
@@ -9,10 +9,10 @@ public class SessionDto {
     private Long id;
     private String sessionName;
     private String customerName;
-    private Long currentPrice;
+    private Double currentPrice;
     private LocalDateTime start;
     private LocalDateTime end;
-    private Long bet;
+    private Double bet;
     private String location;
     private String lastCustomerBet;
     private String status;
